refactor(open-mpm): split mcp/config.rs into config/ submodules (#366)
